Removing all stdout stderr from main library.
