+++ include/simdjson/parsedjson.h
@@ -4,7 +4,6 @@
 #include "simdjson/common_defs.h"
 #include "simdjson/simdjson.h"
 #include <cstring>
-#include <iostream>
 #include <memory>
 
 #define JSON_VALUE_MASK 0xFFFFFFFFFFFFFF
@@ -137,19 +136,6 @@ public:
 
 };
 
-// dump bits low to high
-inline void dumpbits_always(uint64_t v, const std::string &msg) {
-  for (uint32_t i = 0; i < 64; i++) {
-    std::cout << (((v >> static_cast<uint64_t>(i)) & 0x1ULL) ? "1" : "_");
-  }
-  std::cout << " " << msg.c_str() << "\n";
-}
 
-inline void dumpbits32_always(uint32_t v, const std::string &msg) {
-  for (uint32_t i = 0; i < 32; i++) {
-    std::cout << (((v >> i) & 0x1ULL) ? "1" : "_");
-  }
-  std::cout << " " << msg.c_str() << "\n";
-}
 } // namespace simdjson
 #endif
