+++ src/stage2_build_tape.cpp
@@ -1,6 +1,5 @@
 #include <cassert>
 #include <cstring>
-#include <iostream>
 
 #include "simdjson/portability.h"
 #include "simdjson/common_defs.h"
